Seperated mac installation and cleaned up in code
Mac installation is now it's own install step. Added a basic guide to perform a test on a mac.
Cleaned up in all the code snippets in the README. They now have the same formatting. All blocks of code now have three backticks before and after and two tabs (4 spaces).