Changed how bundles are downloaded
Instead of looking for tags in the local git repo we now look in the https://api.github.com/repos/hioa-cs/IncludeOS/releases curl output. This list seems to be ordered with the newest release always at the top. A small change was also made to the python script to allow multiple fields from the json output to be extracted.