Try to include jessie backports if possible
I've tried to add code that will look to add the backports sources for
Jessie so that the current stable Debian release can work out-of-the-box
as well as testing and Ubuntu.

This is not tested as I don't have jessie or Ubuntu installed on my
machine.  It does continue to work correctly on stretch, however.