nix: create a new scope for IncludeOS packages
Instead of manually inheriting stdenv all over, introduce a nixpkgs
overlay and a new package set/scope called pkgsIncludeOS, where stdenv
is set once and automatically used by callPackage.

I created the overlay in a separate file (overlay.nix) in case we want
to re-use it for other `import nixpkgs { ... }` calls.

Since the Nix files currently become part of the input to the IncludeOS
build, this creates a new derivation (build), but the output content is
the same as before (checked with diffoscope).