A nix-shell for OS development
Adds a nix shell for building an IncludeOS bootable, where IncludeOS is
provided as a local cmake installation, for rapid iteration.

1. Build IncludeOS locally, using the default.nix as a shell:
   1. $ nix-shell default.nix
   2. $ mkdir build_includeos && pushd build_includeos
   3. $ cmake .. -DCMAKE_INSTALL_PREFIX:PATH=/home/<you>/includeos_install
   4. $ make -j install
   5. $ popd
2. $ export INCLUDEOS_PACKAGE=/home/<you>/includeos_install
3. $ nix-shell --run 'make -j12 && $vmbuild hello_includeos.elf.bin $bootloader' && file build_example/hello_includeos.elf.bin.img

This should produce a bootable image. You can skip the --run and iterate on the
bootable in the nix shell.

This means there are two nix shells; one for developing the OS and one for the
bootable. This is dumb, but IncludeOS requires some older packages, like old GSL,
which is not a restriction the bootable should need to have. Stil, we should
probably harmonize the two.