Remove GSL
Now that span is a part of the standard, accessible in libc++ as of LLVM 18, the
only need for GSL is to provide Expetcs / Ensures. We used to support the case
when that would throw, which GSL has an option for, but we never want exceptions
thrown in the kernel, so this is only used for testing.

We can revisit the need for a throwing Expects when reviving those tests.