Fix compiler warnings about not packing struct.
When a union member is initialized with a non-trivial value the compiler
complains that the union is not POD-type with respect to layout.

I think the root cause is that POD implies trivial class, and that requires
all default constructors to be trivial. This again requires that
"T has no non-static members with default initializers. (since C++11)".

https://en.cppreference.com/w/cpp/language/default_constructor#Trivial_default_constructor