Change Fixed_list_alloc to allow copying
A change in LLVM's list implementation makes the list copy the allocator in
certain cases, allocating from a temporary copy and deallocating back to the
original. This may be unintentional, but since the standard does not seem to
preclude this use case we need to support it.

Added some test cases that were relevant for finding the issue.