symbols: fix a bug in symbol loading to get backtraces again
This commits adds an __attribute__ to the static struct used to store
ELF metadata to make sure it's included in the ELF .data section.

On boot we first load the symbol table from .elf_symbols and move it to
a new memory location. We then initialise/clear the BSS area. As some of
the variables in the struct were initalised to 0 in the code they would
be cleared again when we initialised BSS and we would drop the new
location of the symbol table.

Also adds more ELF output on boot to help with future debugging.