nix: add optional ccache support
Adds a new attribute withCcache that defaults to false. When enabled, a
wrapper around clang/clang++ will be added to the IncludeOS build
environment.

With an existing cache, this reduces the service build time to about 1/5
on my server when the kernel has to be rebuilt.

The directory used for caching is /nix/var/cache/ccache and it has to
exist with the right permissions. See overlay.nix for additional
details.

To enable ccache, pass "--arg withCcache true" to either nix-build or
nix-shell.

To manage the ccache, install "ccache" locally. Then "ccache --dir
/nix/var/cache/ccache --show-stats" will show cache statistics and
"ccache --dir /nix/var/cache/ccache --clear" will clear the cache.

Ccache is currently disabled on unittests and integration tests to make
sure that these are built from a clean environment.