Build kernel with -fno-threadsafe-statics
Disables automatic locking when static variables are initialised from
constructors.

To delay initialisation of some static variables until the kernel is
ready they are initialised from a constructor. This can for instance be
useful when waiting for BSS to be cleared, RNG to be initialised or heap
to be ready.

As we have thread support in libcxx, the compiler will automatically add
surrounding calls to __cxa_guard_acquire and __cxa_guard_release to make
sure the variable is only set by a single thread/process. While locking
they call syscalls FUTEX and GETTID. As this may happen early in the
boot process (pre-libc) we are not prepared to handle syscalls and the
results may be unexpected. One side effect is that they set errno in
libc which has not been initialised to point to a valid location yet.

With thread safe statics disabled we avoid this issue, but we have to
manually lock in the kernel when SMP is enabled.

Alternatives:

An alternative would be to set _LIBCPP_HAS_NO_THREADS=1. This would
remove thread support - but also features that can be useful later for
SMP support, such as std::lock_guard. We would still have to lock
manually when initialising static variables in IncludeOS' SMP mode.

Bonus: Use C++20 when building libraries