Create default heap PMR allocator to avoid early calls to malloc
Previously we used std::map in memmap, which would proceed to call
libc's malloc() before libc was initialised. This uses std::pmr::map
instead and sets the default PMR allocator to be the same allocator as
we later use in mmap/malloc. This allocator is initialised by the kernel
as part of the heap.

This change also allows the kernel to transparently use other std::pmr
implementations after heap initialisation to avoid malloc calls.