nix: add vmrunner and chainloader to includeos passhtrough
This adds vmrunner and chainloader to the IncludeOS overlay to make it
easier to import them from the shell and unikernels.

Also adds pkgs and pkgsStatic as shortcuts to the pinned nixpkgs +
overlay. This allows some packages that don't require full direct access
to nixpkgs to avoid importing the overlay (such as shell.nix or
example.nix)