Removing arch_prctl from syscall and creating architecture specific musl calls
Set_thread_area and arch_prctl are functions
used for setting the thread pointer. This is used
for thread local storage.

Currently arch_prctl goes through a system call trap.
This patches musl so that it goes through a typical
function call instead.

Also, this moves the implementation of these
function calls to architecture, removing the need for
macros.